Fix printf-like format strings:
- Add required casts so the matching argument is of the expected type.
- Removed unused arguments.
- Added missing arguments.
- Remove some unnecessary calls to `std::string::c_str` - the `string_format` function will do this automatically.

+++ src/overmap.cpp
@@ -4096,7 +4096,7 @@ point om_direction::rotate( const point &p, type dir )
 {
     switch( dir ) {
         case type::invalid:
-            debugmsg( "Invalid overmap rotation (%d).", dir );
+            debugmsg( "Invalid overmap rotation (%d).", static_cast<int>( dir ) );
             // Intentional fallthrough.
         case type::north:
             break;  // No need to do anything.
